pyproject.toml
[project]
name = "ecommerce-eng-dados"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apache-airflow==2.8.1",
    "dbt-core==1.9.0",
    "dbt-postgres==1.9.0",
    "faker>=37.5.3",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
